fix bug

+++ src/builtin.cc
@@ -178,7 +178,7 @@ std::string lower(const std::string &str) {
 
 std::string substr(const std::string &str, const int64_t pos,
                    const int64_t len) {
-  int64_t p = pos;
+  size_t p = pos;
   if (pos < 0) {
     p = str.length() + pos + 1;
   }
@@ -187,7 +187,7 @@ std::string substr(const std::string &str, const int64_t pos,
     return "";
   }
 
-  int64_t n = len;
+  size_t n = len;
   size_t start = p - 1;
   if (n < 0) {
     n = str.length();
